requirements.txt
google-generativeai
telethon
ccxt
python-dotenv